Folder renamed, removed pyc files and added title to README.md
