+++ lib/errors.py
