+++ lib/sdk.py
@@ -1,8 +1,8 @@
-from functions import users
-from functions import things
-from functions import messages
-from functions import channels
-from functions import groups
+from lib import users
+from lib import things
+from lib import messages
+from lib import channels
+from lib import groups
 
 import requests
