+++ tests/test_version.py
@@ -1,4 +1,4 @@
-from functions import sdk
+from lib import sdk
 
 import json
